Добавить рассылку новостей #48

+++ commands/SubscribeController.php
@@ -18,31 +18,36 @@ class SubscribeController extends Controller
      */
     public function actionSend()
     {
-        $time = microtime(true);
-        $list = SubscribeMailItem::query()
-            ->limit(10)
-            ->orderBy(['date_insert' => SORT_DESC])
-            ->all();
-        \Yii::info('Всего писем для рассылки: ' . count($list), 'gs\\app\\commands\\SubscribeController::actionSend');
-        \Yii::info('Список писем: ' . VarDumper::dumpAsString(ArrayHelper::getColumn($list, 'mail')), 'gs\\app\\commands\\SubscribeController::actionSend');
+        $key = 'app\\commands\\SubscribeController::actionSend';
+        $isCache = \Yii::$app->cache->get($key);
+        if ($key !== false) {
+            \Yii::$app->cache->set($key, true, 60*60);
+            $time = microtime(true);
+            $list = SubscribeMailItem::query()
+                ->limit(10)
+                ->orderBy(['date_insert' => SORT_DESC])
+                ->all();
+            \Yii::info('Всего писем для рассылки: ' . count($list), 'gs\\app\\commands\\SubscribeController::actionSend');
+            \Yii::info('Список писем: ' . VarDumper::dumpAsString(ArrayHelper::getColumn($list, 'mail')), 'gs\\app\\commands\\SubscribeController::actionSend');
 
-        foreach($list as $mailItem) {
-            \Yii::$app->mailer
-                ->compose()
-                ->setFrom(\Yii::$app->params['mailer']['from'])
-                ->setTo($mailItem['mail'])
-                ->setSubject($mailItem['subject'])
-                ->setTextBody($mailItem['text'])
-                ->setHtmlBody($mailItem['html'])
-                ->send();
+            foreach($list as $mailItem) {
+                \Yii::$app->mailer
+                    ->compose()
+                    ->setFrom(\Yii::$app->params['mailer']['from'])
+                    ->setTo($mailItem['mail'])
+                    ->setSubject($mailItem['subject'])
+                    ->setTextBody($mailItem['text'])
+                    ->setHtmlBody($mailItem['html'])
+                    ->send();
+            }
+
+            \Yii::info('Список писем для удаления: ' . VarDumper::dumpAsString(ArrayHelper::getColumn($list, 'id')), 'gs\\app\\commands\\SubscribeController::actionSend');
+            SubscribeMailItem::deleteByCondition([
+                'in', 'id', ArrayHelper::getColumn($list, 'id')
+            ]);
+            \Yii::info('Осталось после рассылки: ' . SubscribeMailItem::query()->count(), 'gs\\app\\commands\\SubscribeController::actionSend');
+            \Yii::info('Затраченное время на расылку: ' . microtime(true)-$time, 'gs\\app\\commands\\SubscribeController::actionSend');
         }
-
-        \Yii::info(':Список писем для удаления: ' . VarDumper::dumpAsString(ArrayHelper::getColumn($list, 'id')), 'gs\\app\\commands\\SubscribeController::actionSend');
-        SubscribeMailItem::deleteByCondition([
-            'in', 'id', ArrayHelper::getColumn($list, 'id')
-        ]);
-        \Yii::info('Осталось после рассылки: ' . SubscribeMailItem::query()->count(), 'gs\\app\\commands\\SubscribeController::actionSend');
-        \Yii::info('Затраченное время на расылку: ' . microtime(true)-$time, 'gs\\app\\commands\\SubscribeController::actionSend');
 
         \Yii::$app->end();
     }
